Fix bug where n_subsamples=Inf wasn't properly passed through C++ code

+++ src/RcppExports.cpp
@@ -140,7 +140,7 @@ BEGIN_RCPP
 END_RCPP
 }
 // estimate_overdispersions_fast
-NumericVector estimate_overdispersions_fast(RObject Y, RObject mean_matrix, NumericMatrix model_matrix, bool do_cox_reid_adjustment, int n_subsamples);
+NumericVector estimate_overdispersions_fast(RObject Y, RObject mean_matrix, NumericMatrix model_matrix, bool do_cox_reid_adjustment, double n_subsamples);
 RcppExport SEXP _glmGamPoi_estimate_overdispersions_fast(SEXP YSEXP, SEXP mean_matrixSEXP, SEXP model_matrixSEXP, SEXP do_cox_reid_adjustmentSEXP, SEXP n_subsamplesSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
@@ -149,7 +149,7 @@ BEGIN_RCPP
     Rcpp::traits::input_parameter< RObject >::type mean_matrix(mean_matrixSEXP);
     Rcpp::traits::input_parameter< NumericMatrix >::type model_matrix(model_matrixSEXP);
     Rcpp::traits::input_parameter< bool >::type do_cox_reid_adjustment(do_cox_reid_adjustmentSEXP);
-    Rcpp::traits::input_parameter< int >::type n_subsamples(n_subsamplesSEXP);
+    Rcpp::traits::input_parameter< double >::type n_subsamples(n_subsamplesSEXP);
     rcpp_result_gen = Rcpp::wrap(estimate_overdispersions_fast(Y, mean_matrix, model_matrix, do_cox_reid_adjustment, n_subsamples));
     return rcpp_result_gen;
 END_RCPP
